Fix pointer-to-unsigned-long casts in filter files for UCRT64 (LLP64)
On Windows LLP64 (UCRT64), unsigned long is 32-bit but pointers are
64-bit, so casting SWKey* to unsigned long loses precision and causes
a -fpermissive compile error.

Replace (unsigned long)key with (uintptr_t)key in all affected filter
files and add #include <stdint.h> where needed.

Fixes: latin1utf8.cpp, latin1utf16.cpp, utf8html.cpp, utf8nfc.cpp,
scsuutf8.cpp, utf8latin1.cpp, utf8scsu.cpp, utf8arshaping.cpp,
utf8bidireorder.cpp, utf8nfkd.cpp, cipherfil.cpp

+++ src/modules/filters/cipherfil.cpp
@@ -23,6 +23,7 @@
 
 
 #include <stdlib.h>
+#include <stdint.h>
 #include <cipherfil.h>
 #include <swcipher.h>
 #include <swbuf.h>
@@ -56,7 +57,7 @@ char CipherFilter::processText(SWBuf &text, const SWKey *key, const SWModule *mo
 			text.setSize(len + 5);
 			memcpy(text.getRawData(), cipher->getUncipheredBuf(), len);
 		}
-		else if ((unsigned long)key == 1) {
+		else if ((uintptr_t)key == 1) {
 			cipher->setUncipheredBuf(text.getRawData(), len);
 			cipher->getCipheredBuf(&len);
 			text.setSize(len + 5);
