+++ src/modules/filters/utf8scsu.cpp
@@ -22,6 +22,7 @@
 
 #ifdef _ICU_
 
+#include <stdint.h>
 #include <utf8scsu.h>
 #include <swbuf.h>
 
@@ -42,7 +43,7 @@ UTF8SCSU::~UTF8SCSU() {
 }
 
 char UTF8SCSU::processText(SWBuf &text, const SWKey *key, const SWModule *module) {
-	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((uintptr_t)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
 
 	err = U_ZERO_ERROR;
